requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==2.4.1
asyncer==0.0.2
beautifulsoup4==4.12.3
bidict==0.23.1
cachetools==5.5.0
certifi==2024.12.14
chainlit==1.1.300
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cohere==5.11.4
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.15
distro==1.9.0
executing==2.0.1
fastapi==0.110.3
fastavro==1.9.7
filelock==3.16.1
filetype==1.2.0
fonttools==4.55.0
frontend==0.0.3
fsspec==2024.10.0
google==3.0.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-pubsub==2.27.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
Lazify==0.4.0
literalai==0.0.604
MarkupSafe==3.0.2
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.59.6
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==23.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.2.2
portalocker==2.10.1
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.29.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.25.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.0.3
qdrant-client==1.12.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.45.2
syncer==2.0.3
tokenizers==0.20.3
tomli==2.0.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uptrace==1.24.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==0.20.0
Wave==0.0.2
wcwidth==0.2.13
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
